refactor: Move load_ensemble_nfo, load_models_from_ensemble to utils, update relevant files
